requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.1.0
celery==5.3.4
redis==5.0.1 